fix(ci): python-coverage-shards — split provisioner.py under the 1500-line maintainability gate
This PR's epoch-fencing work pushed src/awf/node/provisioner.py to 1507 lines,
tripping test_first_party_code_files_stay_under_line_limit (limit 1500).

Mechanically extract five self-contained short-transaction helpers
(_verify_execution_claim_epoch, _record_egress_audit_if_current,
_create_egress_audit_record, _issue_secret_leases, _recheck_status) into a new
sibling mixin module provisioner_short_txn_helpers.py, mirroring the existing
ProvisionerHostPortCheckMixin pattern. Behavior is unchanged; the methods are
wired back onto Provisioner via ProvisionerShortTxnHelpersMixin. provisioner.py
drops to 1361 lines.

Update the one test that monkeypatched awf.node.provisioner.SecretLeaseService
to patch the symbol in its new module namespace.
